+++ src/merging/merging.jl
@@ -1,21 +1,3 @@
-mutable struct GridCell
-    np::Int64
-    w::Float64
-    v_mean::SVector{3,Float64}
-    v_std_sq::SVector{3,Float64}
-    x_mean::SVector{3,Float64}
-    x_std_sq::SVector{3,Float64}
-    particle_index1::Int64
-    particle_index2::Int64
-
-    w1::Float64
-    w2::Float64
-    v1::SVector{3,Float64}  # these are for the post-merge quantities
-    v2::SVector{3,Float64}
-    x1::SVector{3,Float64}
-    x2::SVector{3,Float64}
-end
-
 include("merging_grid.jl")
 include("merging_octree_N2.jl")
 include("nnls.jl")