+++ test/test_indexing_particlevector.jl
@@ -1,4 +1,4 @@
-@testset "multidimensional indexing" begin
+@testset "particle vector and multidimensional indexing" begin
     pia = ParticleIndexerArray(8, 3)  # 8 cells 3 species
     @test size(pia.indexer) == (8, 3)
 
@@ -30,4 +30,15 @@
 
         @test particles[1].cell[i] == 0
     end
+
+    old_buffer_length = particles[1].nbuffer
+    resize!(particles[1], length(particles[1]) + 3)
+    @test length(particles[1]) == 13
+    @test length(particles[1].particles) == 13
+    @test length(particles[1].index) == 13
+    @test length(particles[1].cell) == 13
+    @test length(particles[1].buffer) == 13
+
+    @test particles[1].nbuffer == old_buffer_length + 3
+    @test particles[1].buffer == [11,12,13,10,9,8,7,6,5,4,3,2,1]
 end